Make alloc/stackalloc with signed value deprecated. Change usages to always use unsiged value

+++ javalib/src/main/scala/java/io/File.scala
@@ -135,7 +135,7 @@ class File(_path: String) extends Serializable with Comparable[File] {
    * match that of Java on non-existing file.
    */
   private def simplifyExistingPath(path: CString)(implicit z: Zone): CString = {
-    val resolvedName = alloc[Byte](limits.PATH_MAX)
+    val resolvedName = alloc[Byte](limits.PATH_MAX.toUInt)
     realpath(path, resolvedName)
     resolvedName
   }
@@ -345,7 +345,7 @@ object File {
 
   private def getUserDir(): String =
     Zone { implicit z =>
-      var buff: CString = alloc[CChar](4096)
+      var buff: CString = alloc[CChar](4096.toUInt)
       var res: CString  = getcwd(buff, 4095.toUInt)
       fromCString(res)
     }
@@ -479,7 +479,7 @@ object File {
 
   @tailrec private def resolve(path: CString, start: UInt = 0.toUInt)(
       implicit z: Zone): CString = {
-    val part: CString = alloc[Byte](limits.PATH_MAX)
+    val part: CString = alloc[Byte](limits.PATH_MAX.toUInt)
     val `1U`          = 1.toUInt
     // Find the next separator
     var i = start
@@ -512,7 +512,7 @@ object File {
    * Otherwise, returns `None`.
    */
   private def readLink(link: CString)(implicit z: Zone): CString = {
-    val buffer: CString = alloc[Byte](limits.PATH_MAX)
+    val buffer: CString = alloc[Byte](limits.PATH_MAX.toUInt)
     readlink(link, buffer, (limits.PATH_MAX - 1).toUInt) match {
       case -1 =>
         null
